First attempt at correcting orca obstacle collision problems
On the convex circle, I'm intersecting two obstacles at the same time.
One issue is that when an orca agent collides with an obstacle, a constraint
is introduced to keep the agent from penetrating *deeper*.  This isn't
sufficient in the case where the agent is sliding along the obstacle.

This quick hack (on just the ORCA agent) pushes the agent out a bit from
the wall.  The amount is an arbitrary magic number and doesn't use any
reasoned logic.

The second approach is to recognize that I don't need the second obstacle
when I end up colliding with the *next* obstacle.  I need to recognize I'm
colliding with the previous and that my collision is accounted for.  This
is harder but would be the more general solution.